requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
lxml==6.0.0
MarkupSafe==3.0.2
numpy==2.3.1
python-docx==1.2.0
scikit-learn==1.7.0
scipy==1.16.0
threadpoolctl==3.6.0
typing_extensions==4.14.1
Werkzeug==3.1.3
